+++ files/3-rinkhals/start.sh
@@ -11,16 +11,18 @@ quit() {
     echo
     log "/!\\ Startup failed, stopping Rinkhals..."
 
-    B=/sys/class/pwm/pwmchip0/pwm0
-    echo 0 > $B/enable; echo 0 > $B/duty_cycle
-    echo 3817000 > $B/period; echo 1526800 > $B/duty_cycle; echo 1 > $B/enable
-    usleep 300000; echo 0 > $B/enable; usleep 100000
-    echo 0 > $B/duty_cycle
-    echo 4545000 > $B/period; echo 1818000 > $B/duty_cycle; echo 1 > $B/enable
-    usleep 300000; echo 0 > $B/enable; usleep 100000
-    echo 0 > $B/duty_cycle
-    echo 5714000 > $B/period; echo 2285600 > $B/duty_cycle; echo 1 > $B/enable
-    usleep 600000; echo 0 > $B/enable
+    if [ ! -f /useremain/rinkhals/.mute-sounds ]; then
+        B=/sys/class/pwm/pwmchip0/pwm0
+        echo 0 > $B/enable; echo 0 > $B/duty_cycle
+        echo 3817000 > $B/period; echo 1526800 > $B/duty_cycle; echo 1 > $B/enable
+        usleep 300000; echo 0 > $B/enable; usleep 100000
+        echo 0 > $B/duty_cycle
+        echo 4545000 > $B/period; echo 1818000 > $B/duty_cycle; echo 1 > $B/enable
+        usleep 300000; echo 0 > $B/enable; usleep 100000
+        echo 0 > $B/duty_cycle
+        echo 5714000 > $B/period; echo 2285600 > $B/duty_cycle; echo 1 > $B/enable
+        usleep 600000; echo 0 > $B/enable
+    fi
 
     ./stop.sh
     touch /useremain/rinkhals/.disable-rinkhals
@@ -376,4 +378,4 @@ ta_da() {
     echo 1912000 > $B/period; echo 764800 > $B/duty_cycle; echo 1 > $B/enable
     usleep 280000; echo 0 > $B/enable
 }
-ta_da &
+[ -f /useremain/rinkhals/.mute-sounds ] || ta_da &
